requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
assemblyai==0.30.0
audioread==3.1.0
bcrypt==4.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
decorator==5.2.1
Deprecated==1.2.18
dnspython==2.7.0
dotenv==0.9.9
email-validator==2.3.0
fastapi==0.116.1
ffmpeg-python==0.2.0
future==1.0.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.5.2
lazy_loader==0.4
librosa==0.11.0
limits==5.5.0
llvmlite==0.46.0
Mako==1.3.10
MarkupSafe==3.0.2
msgpack==1.1.2
nltk==3.9.1
numba==0.63.1
numpy==2.3.5
packaging==25.0
pandas==2.2.3
passlib==1.7.4
platformdirs==4.5.1
pooch==1.8.2
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyloudnorm==0.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
regex==2025.11.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.43
starlette==0.47.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
uvicorn==0.35.0
websockets==15.0.1
wrapt==1.17.3
